+++ app/components/today/index.ts
@@ -1,6 +1,7 @@
 export * from './BodyWeightCheckIn';
 export * from './CardioLogModal';
 export * from './StreakCard';
+export * from './WeeklyRings';
 export * from './WorkoutPicker';
 export * from './TodayWorkoutCard';
 export * from './StartWorkoutModal';